+++ packages/sitecore-jss-nextjs/src/middleware/middleware.test.ts
@@ -132,11 +132,47 @@ describe('MiddlewareBase', () => {
       expect(middleware['isPrefetch'](req)).to.equal(true);
     });
 
+    it('should return true when x-middleware-prefetch header is 1', () => {
+      const middleware = new SampleMiddleware({ siteResolver: new MockSiteResolver([]) });
+      const req = createReq({
+        headerValues: {
+          'x-middleware-prefetch': '1',
+        },
+      });
+
+      expect(middleware['isPrefetch'](req)).to.equal(true);
+    });
+
     it('should return false when required header is not provided', () => {
       const middleware = new SampleMiddleware({ siteResolver: new MockSiteResolver([]) });
       const req = createReq();
 
       expect(middleware['isPrefetch'](req)).to.equal(false);
+    });
+
+    it('should return false when a mobile device is detected via sec-ch-ua-mobile', () => {
+      const middleware = new SampleMiddleware({ siteResolver: new MockSiteResolver([]) });
+      const req = createReq({
+        headerValues: {
+          purpose: 'prefetch',
+          'sec-ch-ua-mobile': '?1',
+          'x-middleware-prefetch': '1',
+        },
+      });
+
+      expect(middleware['isPrefetch'](req)).to.equal(false);
+    });
+
+    it('should return true when it is a desktop device and purpose is prefetch', () => {
+      const middleware = new SampleMiddleware({ siteResolver: new MockSiteResolver([]) });
+      const req = createReq({
+        headerValues: {
+          purpose: 'prefetch',
+          'sec-ch-ua-mobile': '?0',
+        },
+      });
+
+      expect(middleware['isPrefetch'](req)).to.equal(true);
     });
   });