+++ packages/sitecore-jss-nextjs/src/middleware/personalize-middleware.test.ts
@@ -499,6 +499,35 @@ describe('PersonalizeMiddleware', () => {
     });
   });
 
+  it('should skip personalization for mobile device requests with prefetch headers', async () => {
+    const req = createRequest({
+      headerValues: {
+        purpose: 'prefetch',
+        'sec-ch-ua-mobile': '?1',
+        'x-middleware-prefetch': '1',
+      },
+    });
+    const res = createResponse();
+    const { middleware } = createMiddleware();
+
+    const finalRes = await middleware.getHandler()(req, res);
+
+    validateDebugLog('personalize middleware start: %o', {
+      hostname: 'foo.net',
+      pathname: '/styleguide',
+      language: 'en',
+      headers: {
+        ...req.headers,
+      },
+    });
+
+    expect(debugSpy.args.some((log) => log[0] === 'skipped (prefetch)')).to.equal(
+      false,
+      'Expected debug log "skipped (prefetch)" not to be called'
+    );
+    expect(finalRes).to.deep.equal(res);
+  });
+
   describe('request passed', () => {
     it('fallback defaultLocale is used', async () => {
       const language = 'da-DK';